[MRG] initial refactor of compute command (and associated test module) (#734)
* initial refactor of compute command (and associated tests) into different modules

* ...actually add the files

* localize pathos import to code that uses it, make tests skip if no pysam or pathos

* move pysam import where it is used
