+++ requirements.txt
@@ -1,6 +1,4 @@
 pytest
-pathos
-pysam
 screed
 numpy
 matplotlib
